trying to get negative indirects to work. ld should be now fixed in very, very big quotations

+++ vm/sources/op_zjmp_fork_lfork.c
@@ -21,12 +21,12 @@ void	op_zjmp(t_info *info, t_car *car)
 
 void	op_fork(t_info *info, t_car *car)
 {
-	push_new_car(info, car, (get_argument(info, 1, car) % IDX_MOD));
+	push_new_car(info, car, ((short)get_argument(info, 1, car) % IDX_MOD));
 	car->jump = 3;
 }
 
 void	op_lfork(t_info *info, t_car *car)
 {
-	push_new_car(info, car, get_argument(info, 1, car));
+	push_new_car(info, car, (short)get_argument(info, 1, car));
 	car->jump = 3;
 }
